Add files via upload
The neonate and adult analysis code needed edits to create boarders, properly save the figures, and apply the estimated stretches to the plots. These should be more efficient, still the boarder creation issue is not as nice as it could be, but I do not much more time to work on this. Also uploaded this rather than pushed it from my repository because I was working in a lot of different folders.